adopt_service done

+++ adopt_service/src/main/java/ba/unsa/etf/nwt/adopt_service/controller/AddPetRequestController.java
@@ -1,5 +1,6 @@
 package ba.unsa.etf.nwt.adopt_service.controller;
 
+import ba.unsa.etf.nwt.adopt_service.exception.WrongInputException;
 import ba.unsa.etf.nwt.adopt_service.model.AddPetRequest;
 import ba.unsa.etf.nwt.adopt_service.request.PetForAdoptRequest;
 import ba.unsa.etf.nwt.adopt_service.response.ResponseMessage;
@@ -7,6 +8,8 @@ import ba.unsa.etf.nwt.adopt_service.security.CurrentUser;
 import ba.unsa.etf.nwt.adopt_service.security.UserPrincipal;
 import ba.unsa.etf.nwt.adopt_service.service.AddPetRequestService;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.security.RolesAllowed;
@@ -24,25 +27,43 @@ public class AddPetRequestController {
         return addPetRequestService.getAddPetRequest();
     }
 
-   /* @PostMapping("/add-pet-request")
+    /* @PostMapping("/add-pet-request")
     public ResponseMessage addAddPetRequest(@Valid @RequestBody AddPetRequest addPetRequest) {
         return addPetRequestService.addAddPetRequest(addPetRequest);
     }*/
+
     //izmijenjeni oblik post metode
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
     @PostMapping("/eurekaa/add-pet-request")
     public ResponseMessage addAddPetRequest(@Valid @RequestBody PetForAdoptRequest addPetRequest, @CurrentUser UserPrincipal currentUser) {
        return addPetRequestService.addAddPetRequest(addPetRequest, currentUser);
     }
 
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
     @PostMapping("/add-pet-request")
     public ResponseMessage addAddPetRequestLocal(@Valid @RequestBody AddPetRequest addPetRequest) {
         return addPetRequestService.addAddPetRequestLocal(addPetRequest);
     }
 
-    //trenutni za sebe - admin
-    //todo postaviti standard ako je trenutni user u pitanju onda moze pristupiti...
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
     @GetMapping("/add-pet-request/user/{userID}")
-    public List<AddPetRequest> getAddPetRequestByUserID(@PathVariable Long userID) {
+    public List<AddPetRequest> getAddPetRequestByUserID(@PathVariable Long userID, @CurrentUser UserPrincipal currentUser) {
+
+        //pronalazak role trenutnog korisnika
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        boolean hasAdminRole = authentication.getAuthorities().stream()
+                .anyMatch(r -> r.getAuthority().equals("ROLE_ADMIN"));
+
+        //svi admini i samo useri ciji je request imaju pravo pronalaska istog
+        if(hasAdminRole){
+            return addPetRequestService.getAddPetRequestByUserID(userID);
+        }
+
+        //korisnici mogu pregledati samo vlastite requeste
+        if (!currentUser.getId().equals(userID)) {
+            throw new WrongInputException("Current user is not admin and this request doesn't belong to current user!");
+        }
+
         return addPetRequestService.getAddPetRequestByUserID(userID);
     }
 
@@ -70,10 +91,25 @@ public class AddPetRequestController {
         return addPetRequestService.deleteAddPetRequestByID(id);
     }
 
-    //zast, user samo svoj - admin
-    //todo postaviti standard ako je trenutni user u pitanju onda moze pristupiti...
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
     @DeleteMapping("/add-pet-request/user/{userID}")
-    public ResponseMessage deleteAddPetRequestsByUserID(@PathVariable Long userID) {
+    public ResponseMessage deleteAddPetRequestsByUserID(@PathVariable Long userID, @CurrentUser UserPrincipal currentUser) {
+
+        //pronalazak role trenutnog korisnika
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        boolean hasAdminRole = authentication.getAuthorities().stream()
+                .anyMatch(r -> r.getAuthority().equals("ROLE_ADMIN"));
+
+        //svi admini i samo useri ciji je request imaju pravo brisanja istog
+        if(hasAdminRole){
+            return addPetRequestService.deleteAddPetRequestsByUserID(userID);
+        }
+
+        //korisnici mogu obrisati samo vlastite requeste
+        if (!currentUser.getId().equals(userID)) {
+            throw new WrongInputException("Current user is not admin and this request doesn't belong to current user!");
+        }
+
         return addPetRequestService.deleteAddPetRequestsByUserID(userID);
     }
